Nextcloud 21
Downloaded and unpacked Nextcloud 21

snapshot of psdev data

Added utility for stopping all IDA service components

Added utility for stopping all IDA service components

Added utility for stopping all IDA service components

Removed temporary package files

Removed obsolete files from IDA app. Updated utility script for listing modified files

Updated modification instructions

Updates to modification utility scripts and analysis

Updates to modifications README

Updates to modifications README

Baseline of new version of Nextcloud, ready for modification and debugging

Update to modifications README

Updated scripts for generating modification diffs

Updated scripts for generating modification diffs

Baseline homepage layout OK

Local login working

Tweaks to local login field placeholder text

Updates to modification instructions

Removed obsolete font definitions

Added IDA First Run Wizard app

Updated modification instructions

Working SSO integration

Header and main navigation layout now correct

Removed unwanted options from left hand navigation of Files views

Disabled IDA tab view until it is implemented using new view framework

Completed updates to all modified files in the files app

Completed updates to all modified files in the files_sharing app that seemed relevant

Applied patch for properly handling zero sized files

Completed updates to all nextcloud/core files

Completed updates to all nextcloud/lib/private files

Updates to modification instructions

Completed updates to all nextcloud/lib/public files

Disabled meaningless console error pertaining to avatars

Added all FDWE related code into core templates

Modified user settings app to remove unwanted options

Fixed main navigation header labels

Removed unnecessary test for existing user files when creating new users, as IDA admin scripts and process handle that

Updated project and user admin scripts to force initialization of root /files directory for user accounts

Updated admin and command line tool tests to fully flush all test accounts

Removed undesired elements from sharing tab

Removed undesired options from temporary share link settings menu

Manually  added https://ajax.googleapis.com/ajax/libs/jquery/2.2.4/jquery.min.js  as nextcloud/core/js/jquery.js

IDA tab now loading without errors

Added explicit version number to jquery file

Deleted non-versioned jquery file

IDA tab loading but jquery function calls not working yet

IDA tab now displaying properly and action buttons are active, though still issues executing actions

Added modification utility script for generating new modified diffs

Added missed Swedish translations

Deleted superfluous original files where no modifications were made

Added admin utility for initiating an action with specific user credentials

Fixed jquery errors in frozen node tab

Fixed issues with freezing/unfreezing/deleting files via UI

Fixed dialog button reference, which changed

Fixed incorrect response codes to properly return 403 when privileges are lacking

Fixed missing action status notification icons in header

Tweaks to action notification layout

Removed mobile app QR option from app token settings form

Fixed delete option for staging files and folders

Created admin utility script for checking dataset intersections in Metax

Tweaks to unfreeze/delete tab view button layout

Regenerated new modified file diffs

Added missing use clause for JWT

Purged diffs used during update modifications

Added utility script for remounting dev glusterfs after reboot

Added admin utility scripts for marking actions as cleared, failed, or completed and retrying failed actions

Added X-XSS-Protection header with fixed value of zero

Added check at beginning of admin scripts for presence of OFFLINE sentinel file, causing each script to immediately exit with an error

Fixed config variable name

Added logging of ignored admin requests when service in OFFLINE mode

Updated text on IDA home page

Added new test utility script for initializing one million files, copied from master branch

Updated IDA first run wizard to be compatible with updated version of jquery

Updated method for registering Notifier for IDA first run wizard

Added support for Finnish and Swedish settings left hand navigation section names

Fixed app password creation and management when logged in via SSO

Removed unrequired link behavior from Fairdata IDA header logo

Disabled link behavior from settings view labels

Updated method for registering Notifier

Fixed failure of focus to change to IDA tab when snowflake icon is clicked

All modification diff generation now put in explicit diff subdirectory

Added definition of CLI root to template config

Moved CLI tests from this repository to the CLI tools repository

Disabled collapsible shares navigation submenu

+++ nextcloud/lib/private/Authentication/Token/PublicKeyTokenMapper.php
@@ -1,8 +1,12 @@
 <?php
+
 declare(strict_types=1);
+
 /**
  * @copyright Copyright (c) 2018 Roeland Jago Douma <roeland@famdouma.nl>
  *
+ * @author Christoph Wurst <christoph@winzerhof-wurst.at>
+ * @author Daniel Kesselberg <mail@danielkesselberg.de>
  * @author Roeland Jago Douma <roeland@famdouma.nl>
  *
  * @license GNU AGPL version 3 or any later version
@@ -18,7 +22,7 @@ declare(strict_types=1);
  * GNU Affero General Public License for more details.
  *
  * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
 
@@ -29,8 +33,10 @@ use OCP\AppFramework\Db\QBMapper;
 use OCP\DB\QueryBuilder\IQueryBuilder;
 use OCP\IDBConnection;
 
+/**
+ * @template-extends QBMapper<PublicKeyToken>
+ */
 class PublicKeyTokenMapper extends QBMapper {
-
 	public function __construct(IDBConnection $db) {
 		parent::__construct($db, 'authtoken');
 	}
@@ -163,7 +169,8 @@ class PublicKeyTokenMapper extends QBMapper {
 		$qb = $this->db->getQueryBuilder();
 
 		$qb->delete('authtoken')
-			->where($qb->expr()->eq('type', $qb->createNamedParameter(IToken::TEMPORARY_TOKEN)))
+			->where($qb->expr()->eq('uid', $qb->createNamedParameter($except->getUID())))
+			->andWhere($qb->expr()->eq('type', $qb->createNamedParameter(IToken::TEMPORARY_TOKEN)))
 			->andWhere($qb->expr()->neq('id', $qb->createNamedParameter($except->getId())))
 			->andWhere($qb->expr()->eq('version', $qb->createNamedParameter(PublicKeyToken::VERSION, IQueryBuilder::PARAM_INT)));
